# EVM now skips locals init