Improves a `VirtualMachine` execution for a simple test case with a few operations over `Uint256` by 4%, introducing `[SkipLocalsInit]` on the execution of the EVM. EVM has a large stack (over 200 variables) and cleaning, according to benchmarks may take a while. Of course this is a one time setup for execution, the longer it runs, the less is saved by stack init.

## Changes:
- `[SkipLocalsInit]`  + a benchmark

## Types of changes

What types of changes does your code introduce?
_Put an `x` in the boxes that apply_

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Other (please describe): 

## Testing
**Requires testing**

- [ ] Yes
- [x] No

**In case you checked yes, did you write tests??**

- [ ] Yes
- [x] No


## Benchmarks 

### Before

|             Method |        Mean |     Error |    StdDev | Ratio | RatioSD |
|------------------- |------------:|----------:|----------:|------:|--------:|
|        ExecuteCode | 1,892.39 ns | 24.949 ns | 23.338 ns | 25.82 |    0.45 |
| No_machine_running |    73.29 ns |  1.206 ns |  1.128 ns |  1.00 |    0.00 |

### After

|             Method |        Mean |     Error |     StdDev | Ratio | RatioSD |
|------------------- |------------:|----------:|-----------:|------:|--------:|
|        ExecuteCode | 2,035.04 ns | 47.319 ns | 136.526 ns | 24.25 |    1.37 |
| No_machine_running |    82.94 ns |  1.657 ns |   1.469 ns |  1.00 |    0.00