@Scooletz 2022-04-01
It's not. That's why provided the baseline to show the ratio. This is the only applicable finding out of the mentioned #3907 . If seems unrelevant or too flaky, let's close.